refactor(headers)!: move header files to project-specific directory
