+++ include/backtest-cpp/data.h
@@ -4,7 +4,7 @@
 #include <map>
 #include <vector>
 
-#include "types.h"
+#include "backtest-cpp/types.h"
 
 class DataHandler {
    public:
