+++ include/backtest-cpp/performance.h
