+++ src/utils.cpp
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "backtest-cpp/utils.h"
 
 #include <algorithm>
 #include <filesystem>
